requirements.txt
opencv-python
numpy
pyserial
ultralytics   # optional (YOLOv8)